
\name{initDB}
\alias{initDB}
\title{Data Setup}
\usage{
}
\arguments{
  \item{Username}{A character string with the username when you need to access to server, i.e., ORE, Oracle and SQL. The default value is "".}
  
  \item{Password}{A character string with the password when you need to access to server, i.e., ORE, Oracle and SQL. The default value is "".}
  
  \item{Conn_String}{A character string with the conn_string when you need to access to ORE server. The default value is "".}
  
  \item{Database}{A character string with the database, you need to choose \code{public} or \code{private} when your use flat data. The default value is "", means \code{public}.}

}
\description{
  Setup the information to access the database.
}
\details{
}



