% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{initwk}
\alias{initwk}
\title{Create the UI to establish the current working directory}
\usage{
initwk()
}
\value{
None
}
\description{
Create the UI to establish the current working directory
}
