% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{is.charv}
\alias{is.charv}
\title{Test x is identical to empty string}
\usage{
is.charv(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Returns boolean
}
\description{
Test x is identical to empty string
}
