% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{is.charv2}
\alias{is.charv2}
\title{Test x is identical to NA}
\usage{
is.charv2(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Returns boolean
}
\description{
Test x is identical to NA
}
