% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{is.charv3}
\alias{is.charv3}
\title{Test x is identical to char(0)}
\usage{
is.charv3(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Returns boolean
}
\description{
Test x is identical to char(0)
}
