% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{is.charv4}
\alias{is.charv4}
\title{Test x is identical to NULL}
\usage{
is.charv4(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Returns boolean
}
\description{
Test x is identical to NULL
}
