% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{is.charv5}
\alias{is.charv5}
\title{Test x is not equal to TRUE, FALSE, Union or Intersection}
\usage{
is.charv5(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Returns boolean
}
\description{
Test x is not equal to TRUE, FALSE, Union or Intersection
}
