% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{is.charv6}
\alias{is.charv6}
\title{Test x is identical to int0}
\usage{
is.charv6(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Returns boolean
}
\description{
Test x is identical to int0
}
