% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{loadDB}
\alias{loadDB}
\title{Automatically load a passed saved database profile in the research folder}
\usage{
loadDB()
}
\value{
None
}
\description{
Automatically load a passed saved database profile in the research folder
}
