% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{QuotationMarkAdd}
\alias{QuotationMarkAdd}
\title{Build string with quote marks}
\usage{
QuotationMarkAdd(strings)
}
\arguments{
\item{strings}{A string to add quotation marks to.}
}
\value{
Returns input string collapsed with quotemarks
}
\description{
Build string with quote marks
}
