% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{saveDB}
\alias{saveDB}
\title{Save the current database profile in the research folder}
\usage{
saveDB()
}
\value{
None
}
\description{
Save the current database profile in the research folder
}
