% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{saveJob}
\alias{saveJob}
\title{Writes a file intended to save the execution string to disk}
\usage{
saveJob(filename, textString)
}
\arguments{
\item{filename}{name of file to write}

\item{textString}{string of R commands to write}
}
\value{
None
}
\description{
Writes a file intended to save the execution string to disk
}
