% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{setrschfld}
\alias{setrschfld}
\title{Display the UI to set the research base directory}
\usage{
setrschfld()
}
\value{
None
}
\description{
Display the UI to set the research base directory
}
