% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.RDataXMan.R
\name{setupwizard}
\alias{setupwizard}
\title{Display the wizard to identify the users data file/database connection}
\usage{
setupwizard()
}
\value{
None
}
\description{
Display the wizard to identify the users data file/database connection
}
